// src/features/source/pane/types.ts
export interface ConfigWithDepth {
  name: string
  isActive?: boolean
  parentConfiguration?: string | null
  tabNumber?: string
  description?: string
  depth: number
}